added product count and summ

+++ app/Http/Controllers/BasketController.php
@@ -29,7 +29,15 @@ class BasketController extends Controller
 		} else {
 			$order = Order::find($orderId);	
 		}
-		$order ->products()->attach($productId);
+		
+		if ($order->products->contains($productId))
+		{
+			$pivotRow = $order->products()->where('product_id', $productId)->first()->pivot;
+			$pivotRow->count++;
+			$pivotRow->update();					
+		} else {
+				$order ->products()->attach($productId);
+		  }		
 		
 		return redirect() ->route('basket-place');
 		
@@ -42,8 +50,23 @@ class BasketController extends Controller
 			return redirect() ->route('basket-place');		
 		}
 		
-		$order = Order::find($orderId);		
-		$order->products()->detach($productId);		 
+		$order = Order::find($orderId);
+		
+		if ($order->products->contains($productId))
+		{
+			$pivotRow = $order->products()->where('product_id', $productId)->first()->pivot;
+			
+			if ($pivotRow->count < 2)
+			{
+				$order->products()->detach($productId);
+			}
+				else
+				{
+					$pivotRow->count--;
+					$pivotRow->update();
+				}					
+		}	
+		
 		return redirect() ->route('basket-place');
 	}	
 }
